Fix two major regressions in stereo-image generation.
Previously the eyes were swapped, and also the depthmap was applied incorrectly unless "invert depthmap" option was enabled. Now the eyes position should be correct, and "invert depthmap" option is not needed and disabled (if stereo images are generated) to avoid confusion.

I am very sorry to have caused this regression.